신청한 결재 불러오기 기능 구현중

+++ src/main/java/com/gobit/minipj_gobit/service/ApprovalService.java
@@ -5,6 +5,6 @@ import com.gobit.minipj_gobit.entity.Approval;
 public interface ApprovalService {
     void saveApproval(Approval approval);
 
-    Approval getApproval(Integer appNum);
+    Approval getApproval(long appNum);
     
 }
